+++ internal/observability/extractor.go
@@ -18,53 +18,20 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package observerware
+package observability
 
 import (
 	"context"
-	"time"
 
-	"go.uber.org/yarpc/api/transport"
+	"go.uber.org/zap"
+	"go.uber.org/zap/zapcore"
 )
 
-type fakeAck struct{}
+// A ContextExtractor pulls any relevant request-scoped data (e.g., tracing
+// spans) from the request's Context.
+type ContextExtractor func(context.Context) zapcore.Field
 
-func (a fakeAck) String() string { return "" }
-
-type fakeHandler struct {
-	err error
-}
-
-func (h fakeHandler) Handle(_ context.Context, _ *transport.Request, _ transport.ResponseWriter) error {
-	return h.err
-}
-
-func (h fakeHandler) HandleOneway(_ context.Context, _ *transport.Request) error {
-	return h.err
-}
-
-type fakeOutbound struct {
-	transport.Outbound
-
-	err error
-}
-
-func (o fakeOutbound) Call(_ context.Context, _ *transport.Request) (*transport.Response, error) {
-	if o.err != nil {
-		return nil, o.err
-	}
-	return &transport.Response{}, nil
-}
-
-func (o fakeOutbound) CallOneway(_ context.Context, _ *transport.Request) (transport.Ack, error) {
-	if o.err != nil {
-		return nil, o.err
-	}
-	return fakeAck{}, nil
-}
-
-func stubTime() func() {
-	prev := _timeNow
-	_timeNow = func() time.Time { return time.Time{} }
-	return func() { _timeNow = prev }
+// NewNopContextExtractor returns a no-op ContextExtractor.
+func NewNopContextExtractor() ContextExtractor {
+	return ContextExtractor(func(_ context.Context) zapcore.Field { return zap.Skip() })
 }